+++ tests/test_pdfbase_ttfonts.py
@@ -1,24 +1,37 @@
-
 """Test TrueType font subsetting & embedding code.
 
 This test uses a sample font (Vera.ttf) taken from Bitstream which is called Vera
 Serif Regular and is covered under the license in ../fonts/bitstream-vera-license.txt.
 """
-from reportlab.lib.testutils import setOutDir,makeSuiteForClasses, outputfile, printLocation, NearTestCase
+from reportlab.lib.testutils import setOutDir,makeSuiteForClasses, outputfile, printLocation, NearTestCase, rlSkipUnless
 if __name__=='__main__':
     setOutDir(__name__)
-import unittest
+import unittest, os
 from io import BytesIO
 from reportlab.pdfgen.canvas import Canvas
+from reportlab.pdfgen.textobject import PDFTextObject
 from reportlab.pdfbase import pdfmetrics
 from reportlab.pdfbase.pdfdoc import PDFDocument, PDFError
 from reportlab.pdfbase.ttfonts import TTFont, TTFontFace, TTFontFile, TTFOpenFile, \
                                       TTFontParser, TTFontMaker, TTFError, \
                                       makeToUnicodeCMap, \
                                       FF_SYMBOLIC, FF_NONSYMBOLIC, \
-                                      calcChecksum, add32
+                                      calcChecksum, add32, uharfbuzz, shapeFragWord, \
+                                      ShapedFragWord, ShapedStr, ShapeData, _sdSimple, \
+                                      freshTTFont
+from reportlab.platypus.paragraph import Paragraph, _HSFrag
+from reportlab.lib.styles import getSampleStyleSheet
+import zlib, base64
 from reportlab import rl_config
 from reportlab.lib.utils import int2Byte
+from reportlab.lib.abag import ABag
+
+try:
+    TTFont("DejaVuSans", "DejaVuSans.ttf")
+except:
+    haveDejaVuSans = False
+else:
+    haveDejaVuSans = True
 
 def utf8(code):
     "Convert a given UCS character index into UTF-8"
@@ -90,11 +103,8 @@ class TTFontsTestCase(unittest.TestCase):
         "Test PDF generation with TrueType fonts"
         pdfmetrics.registerFont(TTFont("Vera", "Vera.ttf"))
         pdfmetrics.registerFont(TTFont("VeraBI", "VeraBI.ttf"))
-        try:
-            pdfmetrics.registerFont(TTFont("DejaVuSans", "DejaVuSans.ttf"))
-        except:
-            pass
-        else:
+        if haveDejaVuSans:
+            pdfmetrics.registerFont(freshTTFont('DejaVuSans', 'DejaVuSans.ttf'))
             show_all_glyphs('test_pdfbase_ttfonts_dejavusans.pdf',fontName='DejaVuSans')
         show_all_glyphs('test_pdfbase_ttfonts_vera.pdf',fontName='Vera')
         _simple_subset_generation('test_pdfbase_ttfonts1.pdf',1)
@@ -253,7 +263,6 @@ class TTFontFaceTestCase(unittest.TestCase):
         fontFile = doc.idToObject[fontFile.name]
         self.assertTrue(fontFile.content != "")
 
-
 class TTFontTestCase(NearTestCase):
     "Tests TTFont class"
 
@@ -271,7 +280,6 @@ class TTFontTestCase(NearTestCase):
         # Glyphless variation of vedaal's invisible font retrieved from
         # http://www.angelfire.com/pr/pgpf/if.html, which says:
         # 'Invisible font' is unrestricted freeware. Enjoy, Improve, Distribute freely
-        import zlib, base64
         font = """
         eJzdlk1sG0UUx/+zs3btNEmrUKpCPxikSqRS4jpfFURUagmkEQQoiRXgAl07Y3vL2mvt2ml8APXG
         hQPiUEGEVDhWVHyIC1REPSAhBOWA+BCgSoULUqsKcWhVBKjhzfPU+VCi3Flrdn7vzZv33ryZ3TUE
@@ -443,6 +451,243 @@ CMapName currentdict /CMap defineresource pop
 end
 end""")
 
+    @rlSkipUnless(uharfbuzz,'no harfbuzz support')
+    def test_ShapedStrOps(self):
+        sd = ShapedStr('0123456789',shapeData=list(range(10)))
+        e = []
+        def sdcmp(op,v,ex):
+            e.append(f'{op} errors:')
+            i=len(e)
+            if v!=ex: e.append(f'{v!a}!={ex!a}')
+            if type(v)!=type(ex):
+                e.append(f'{v.__class__.__name__}!={ex.__class__.__name__}')
+            elif isinstance(ex,ShapedStr) and v.__shapeData__!=ex.__shapeData__:
+                e.append(f'{v.__shapeData__}!={ex.__shapeData__}')
+            if len(e)==i:
+                e.pop()
+            else:
+                e[i-1] = ' '.join(e[i-1:])
+                del e[i:]
+        sdcmp('sd[1]',sd[1],ShapedStr('1',[1]))
+        sdcmp('sd[2]',sd[2],ShapedStr('2',[2]))
+        sdcmp('sd[-1]',sd[-1],ShapedStr('9',[9]))
+        sdcmp('sd[:2]',sd[:2],ShapedStr('01',[0,1]))
+        sdcmp('sd[:-8]',sd[:-8],ShapedStr('01',[0,1]))
+        sdcmp('sd[-3:-2]',sd[-3:-2],ShapedStr('7',[7]))
+        sdcmp('sd[:0]',sd[:0],'')
+        sdcmp('sd[1:1]',sd[1:1],'')
+        sdcmp('sd[-2:-2]',sd[-2:-2],'')
+        sdcmp('sd[:2]+sd[-2:]',sd[:2]+sd[-2:],ShapedStr('0189',[0,1,8,9]))
+        sdcmp('sd+"ABC"',sd+"ABC",ShapedStr('0123456789ABC',list(range(10))+3*[_sdSimple]))
+        sdcmp('"ABC"+sd',"ABC"+sd,ShapedStr('ABC0123456789',3*[_sdSimple]+list(range(10))))
+        sd += 'ABC'
+        sdcmp('sd+="ABC"',sd,ShapedStr('0123456789ABC',list(range(10))+3*[_sdSimple]))
+        e = '\n'.join(e)
+        self.assertEqual('',e)
+
+    def hbIfaceTest(self, ttfpath, text, exLen, exText, exShapedData):
+        ttfn = os.path.splitext(os.path.basename(ttfpath))[0]
+        ttf = freshTTFont(ttfn, ttfpath)
+        fontName = ttf.fontName
+        fontSize = 30
+        pdfmetrics.registerFont(ttf)
+        w = [pdfmetrics.stringWidth(text,fontName,fontSize),(ABag(fontName=fontName,fontSize=fontSize),text)]
+        new = shapeFragWord(w)
+        ttf.unregister()
+        self.assertEqual(len(new),2,'expected a len of 2')
+        self.assertTrue(isinstance(new,ShapedFragWord),f'returned list class is {new.__class__.__name__} not expected ShapedFragWord')
+        self.assertEqual(new[0],exLen,f'new[0]={new[0]} not expected ={exLen}')
+        self.assertTrue(isinstance(new[1][1],ShapedStr),f'returned str class is {new[1].__class__.__name__} not expected ShapedStr')
+        self.assertTrue(new[1][1]==exText,'shaped string is wrong')
+        self.assertEqual(new[1][1].__shapeData__,exShapedData, 'shape data is wrong')
+
+    @rlSkipUnless(uharfbuzz,'no harfbuzz support')
+    def test_hb_shape_change(self):
+        return self.hbIfaceTest('hb-test.ttf','\u1786\u17D2\u1793\u17B6\u17C6|', 44.22,'\ue000\ue001\u17c6|',
+                                [ShapeData(cluster=0, x_advance=923, y_advance=0, x_offset=0, y_offset=0, width=923),
+                                ShapeData(cluster=0, x_advance=0, y_advance=0, x_offset=-296, y_offset=-26, width=0),
+                                ShapeData(cluster=4, x_advance=0, y_advance=0, x_offset=47, y_offset=-29, width=0),
+                                ShapeData(cluster=5, x_advance=551, y_advance=0, x_offset=0, y_offset=0, width=551)])
+
+    @rlSkipUnless(uharfbuzz,'no harfbuzz support')
+    def test_hb_ligature(self):
+        #ligatures cause the standard length 133.2275390625 to be reduced to 130.78125
+        return self.hbIfaceTest('Vera.ttf','Aon Way',130.78125,'Aon Way',
+                                [ShapeData(cluster=0, x_advance=675.29296875, y_advance=0, x_offset=0.0, y_offset=0, width=684.08203125),
+                                ShapeData(cluster=1, x_advance=603.02734375, y_advance=0, x_offset=-8.7890625, y_offset=0, width=611.81640625),
+                                ShapeData(cluster=2, x_advance=633.7890625, y_advance=0, x_offset=0.0, y_offset=0, width=633.7890625),
+                                ShapeData(cluster=3, x_advance=317.87109375, y_advance=0, x_offset=0.0, y_offset=0, width=317.87109375),
+                                ShapeData(cluster=4, x_advance=956.54296875, y_advance=0, x_offset=0.0, y_offset=0, width=988.76953125),
+                                ShapeData(cluster=5, x_advance=581.0546875, y_advance=0, x_offset=-31.73828125, y_offset=0, width=612.79296875),
+                                ShapeData(cluster=6, x_advance=591.796875, y_advance=0, x_offset=0.0, y_offset=0, width=591.796875)])
+
+
+    @staticmethod
+    def drawVLines(canv,x,y,fontSize,X):
+        canv.saveState()
+        canv.setLineWidth(0.5)
+        canv.setStrokeColor((1,0,0))
+        canv.lines([(_+x,y-0.2*fontSize,_+x,y+fontSize) for _ in X])
+        canv.restoreState()
+
+    @classmethod
+    def drawParaVLines(cls,canv,x,y,p,ttfn):
+        X = [0]
+        mfs = 0
+        frags = p.frags
+        for f in frags:
+            for t in f[1:]:
+                fontName = t[0].fontName
+                if fontName!=ttfn: continue
+                fontSize = t[0].fontSize
+                mfs = max(mfs,fontSize)
+                v = t[1]
+                if type(v) is str:
+                    for s in v:
+                        X.append(X[-1]+pdfmetrics.stringWidth(s,fontName,fontSize))
+                else:
+                    for d in v.__shapeData__:
+                        X.append(X[-1]+d.x_advance*fontSize/1000)
+            if fontName!=ttfn: break
+            if isinstance(f,_HSFrag):
+                X.append(X[-1]+pdfmetrics.stringWidth(' ',fontName,fontSize))
+        if mfs: cls.drawVLines(canv,x,y,mfs,X)
+
+    @staticmethod
+    def shapedStrAdvances(s,fontSize):
+        A = [0]
+        for a in s.__shapeData__:
+            A.append(fontSize*a.x_advance/1000 + A[-1])
+        return A
+
+    @rlSkipUnless(uharfbuzz,'no harfbuzz support')
+    def test_hb_examples(self):
+        canv = Canvas(outputfile('test_pdfbase_ttfonts_hb_examples.pdf'))
+        def hb_example(ttfpath, text, y, fpdfLiteral, excode):
+            ttfn = os.path.splitext(os.path.basename(ttfpath))[0]
+            ttf = freshTTFont(ttfn, ttfpath, shaped=False)
+            try:
+                fontName = ttf.fontName
+                fontSize = 30
+                pdfmetrics.registerFont(ttf)
+                ttf.splitString(text,canv._doc)
+                w = [pdfmetrics.stringWidth(text,fontName,fontSize),(ABag(fontName=fontName,fontSize=fontSize),text)]
+                new = shapeFragWord(w)
+                advances = self.shapedStrAdvances(new[1][1],fontSize)
+                canv.setFont('Helvetica',12)
+                x1 = 400
+                canv.drawString(x1,y,f'unshaped {fontName}')
+                canv.drawString(x1,y-1.2*fontSize,f'fpdf2 shaping {fontName}')
+                canv.drawString(x1,y-2*1.2*fontSize,f'rl shaping {fontName}')
+                canv.setFont(fontName,fontSize)
+                canv.drawString(36,y,text)
+                self.drawVLines(canv,36,y,fontSize,
+                   [pdfmetrics.stringWidth(text[:_],fontName,fontSize) for _ in range(len(text)+1)])
+                canv.saveState()
+                canv.translate(0,-fontSize*1.2)
+                ttf.isShaped = True
+                t = canv.beginText(36, y-1.2*fontSize) #786 - 1.2*30
+                t._textOut(new[1][1],False)
+                code = t.getCode()
+                canv.drawText(t)
+                self.drawVLines(canv,36,y-1.2*fontSize,fontSize,advances)
+                if fpdfLiteral:
+                    canv.addLiteral(fpdfLiteral)
+                    self.drawVLines(canv,36,y,fontSize,advances)
+                canv.restoreState()
+                if code!=excode:
+                    canv.showPage()
+                    canv.save()
+                self.assertEqual(code,excode,f'{fontName} PDF _textOut is wrong\n')
+            finally:
+                ttf.unregister()
+        hb_example(
+            'Vera.ttf',
+            'Aon Way',
+            786,
+            fpdfLiteral='''BT 1 0 0 1 36 786 Tm /F2+0 30 Tf 36 TL (A) Tj 1 0 0 1 56 786 Tm (o) Tj 1 0 0 1 74.35 786 Tm (n) Tj 1 0 0 1 93.36 786 Tm ( ) Tj 1 0 0 1 102.9 786 Tm (W) Tj 1 0 0 1 130.64 786 Tm (a) Tj 1 0 0 1 149.03 786 Tm (y) Tj 1 0 0 1 166.78 786 Tm ET''',
+            excode='''BT 1 0 0 1 36 750 Tm /F2+0 30 Tf 36 TL [(A) 17.57812 (on W) 63.96484 (ay)] TJ ET'''
+            )
+        canv.translate(0,-36)
+        hb_example(
+            'hb-test.ttf',
+            '\u1786\u17D2\u1793\u17B6\u17C6|',
+            706,
+            fpdfLiteral=r'''BT 1 0 0 1 0 0 Tm 1 0 0 1 36 706 Tm /F3+0 30 Tf 36 TL (\006) Tj 1 0 0 1 54.81 705.22 Tm (\007) Tj 1 0 0 1 63.69 706 Tm 1 0 0 1 65.1 705.13 Tm (\005) Tj 1 0 0 1 63.49 706 Tm (|) Tj  ET''',
+            excode = r'''BT 1 0 0 1 36 670 Tm /F3+0 30 Tf 36 TL (\006) Tj -0.78 Ts [296 (\007) -296] TJ -0.87 Ts [-47 (\005) 47 (|)] TJ ET'''
+            )
+        canv.translate(0,-50)
+        if haveDejaVuSans:
+            hb_example(
+                'DejaVuSans.ttf',
+                'Huffing Clifftop finish|',
+                706-70,
+                fpdfLiteral=r'''BT 1 0 0 1 36 636 Tm /F4+0 30 Tf 36 TL (H) Tj 1 0 0 1 58.56 636 Tm (u) Tj 1 0 0 1 77.57 636 Tm (\001) Tj 1 0 0 1 106.58 636 Tm (n) Tj 1 0 0 1 125.59 636 Tm (g) Tj 1 0 0 1 144.63 636 Tm ( ) Tj 1 0 0 1 154.17 636 Tm (C) Tj 1 0 0 1 175.12 636 Tm (l) Tj 1 0 0 1 183.45 636 Tm (i) Tj 1 0 0 1 191.79 636 Tm (\002) Tj 1 0 0 1 212.46 636 Tm (t) Tj 1 0 0 1 224.22 636 Tm (o) Tj 1 0 0 1 242.57 636 Tm (p) Tj 1 0 0 1 261.62 636 Tm ( ) Tj 1 0 0 1 271.15 636 Tm (\003) Tj 1 0 0 1 290.05 636 Tm (n) Tj 1 0 0 1 309.06 636 Tm (i) Tj 1 0 0 1 317.4 636 Tm (s) Tj 1 0 0 1 333.03 636 Tm (h) Tj 1 0 0 1 352.04 636 Tm (|) Tj 1 0 0 1 362.15 636 Tm ET''',
+                excode=r'''BT 1 0 0 1 36 600 Tm /F4+0 30 Tf 36 TL (Hu\001ng Cli\002top \003nish|) Tj ET'''
+                )
+
+        canv.showPage()
+        canv.save()
+
+    def hb_paragraph_drawon(self, canv, ttfpath, text, y=786, textp=None):
+        try:
+            ttfn = os.path.splitext(os.path.basename(ttfpath))[0]
+            ttfn1 = ttfn+'1'
+            ttf = freshTTFont(ttfn,ttfpath,shaped=False)
+            ttf1 = freshTTFont(ttfn1,ttfpath,shaped=True)
+            ttf1.face.name = ttf.face.name + b'1'
+            pdfmetrics.registerFont(ttf)
+            pdfmetrics.registerFont(ttf1)
+            stysh = getSampleStyleSheet()
+            fontSize = 30
+            leading = 36
+            sty = stysh.Normal.clone('sty',fontName=ttfn,fontSize=fontSize,leading=leading)
+            sty1 = stysh.Normal.clone('sty1',fontName=ttfn1,fontSize=fontSize,leading=leading)
+            pW, pH = canv._pagesize
+            x = 36
+            aW = pW - 2*36
+            aH = pH - 2*36
+            p = Paragraph(f'{text}<span face="Helvetica" size="12"> {ttfn} unshaped</span>',sty)
+            w,h = p.wrap(aW,aH)
+            p.drawOn(canv, x=x, y=y)
+            self.drawParaVLines(canv,x,y,p,ttfn)
+            y -= leading
+            p1 = Paragraph(f'{text}<span face="Helvetica" size="12"> {ttfn} shaped</span>',sty1)
+            w,h = p1.wrap(aW,aH)
+            p1.drawOn(canv, x=x, y=y)
+            self.drawParaVLines(canv,x,y,p1,ttfn1)
+            y -= 12
+            xe = pW - 36
+            xm = pW*0.5
+            canv.saveState()
+            if textp is None: textp = text
+            canv.setFont(ttfn,10)
+            canv.drawString(x,y,textp)
+            canv.drawRightString(xe,y,textp)
+            canv.drawCentredString(xm,y,textp)
+            y -= 12
+            canv.setFont(ttfn1,10)
+            canv.drawString(x,y,textp)
+            canv.drawRightString(xe,y,textp)
+            canv.drawCentredString(xm,y,textp)
+            canv.restoreState()
+            y -= 12
+        finally:
+            for _ in ('ttf','ttf1'):
+                if _ in locals():
+                    locals()[_].unregister()
+
+    @rlSkipUnless(uharfbuzz,'no harfbuzz support')
+    def test_hb_paragraph_drawOn(self):
+        canv = Canvas(outputfile('test_pdfbase_ttfonts_hb_para_drawOn.pdf'))
+        self.hb_paragraph_drawon(canv, 'hb-test.ttf', '\u1786\u17D2\u1793\u17B6\u17C6|', y=786)
+        self.hb_paragraph_drawon(canv, 'Vera.ttf', 'Aon Way|', y=786-2*48)
+        if haveDejaVuSans:
+            self.hb_paragraph_drawon(canv, 'DejaVuSans.ttf',
+                            '<span color="green">H</span>u<span color="blue">f</span>fing <u>Clifftop</u> <a href="https://www.reportlab.com">finish</a>|',
+                            y=786-4*48, textp='Huffing Clifftop finish|')
+        canv.showPage()
+        canv.save()
 
 def makeSuite():
     suite = makeSuiteForClasses(